requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
cssselect2==0.7.0
fastapi==0.115.6
fonttools==4.55.3
h11==0.14.0
idna==3.10
passlib==1.7.4
pillow==11.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
python-decouple==3.8
sniffio==1.3.1
starlette==0.41.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
uvicorn==0.34.0
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
